feat: add theme preset and token sync hooks
