+++ apps/cms/src/app/cms/shop/[shop]/themes/useThemeEditor.ts
@@ -1,10 +1,9 @@
 // apps/cms/src/app/cms/shop/[shop]/themes/useThemeEditor.ts
 "use client";
-import { useState, useMemo, useRef, useEffect, type ChangeEvent } from "react";
-import { patchShopTheme } from "../../../wizard/services/patchTheme";
+import { useState, useMemo, useRef } from "react";
 import { tokenGroups } from "./tokenGroups";
-import { useThemePresets } from "./useThemePresets";
-import { savePreviewTokens } from "../../../wizard/previewTokens";
+import { useThemePresetManager } from "./useThemePresetManager";
+import { useThemeTokenSync } from "./useThemeTokenSync";
 
 interface Options {
   shop: string;
@@ -23,13 +22,11 @@ export function useThemeEditor({
   initialOverrides,
   presets,
 }: Options) {
-  const [theme, setTheme] = useState(initialTheme);
-  const [overrides, setOverrides] =
-    useState<Record<string, string>>(initialOverrides);
-  const [, setThemeDefaults] = useState<Record<string, string>>(
-    tokensByTheme[initialTheme]
-  );
   const {
+    theme,
+    setTheme,
+    overrides,
+    setOverrides,
     availableThemes,
     tokensByThemeState,
     presetThemes,
@@ -37,51 +34,48 @@ export function useThemeEditor({
     setPresetName,
     handleSavePreset,
     handleDeletePreset,
-  } = useThemePresets({
+    handleThemeChange,
+  } = useThemePresetManager({
     shop,
-    initialThemes: themes,
-    initialTokensByTheme: tokensByTheme,
+    themes,
+    tokensByTheme,
+    initialTheme,
+    initialOverrides,
     presets,
+  });
+
+  const {
+    previewTokens,
+    mergedTokens,
+    handleOverrideChange,
+    handleReset,
+    handleGroupReset,
+    handleResetAll,
+  } = useThemeTokenSync({
+    shop,
     theme,
     overrides,
-    setTheme,
+    tokensByThemeState,
     setOverrides,
-    setThemeDefaults,
   });
-  const [contrastWarnings, setContrastWarnings] = useState<
-    Record<string, string>
-  >({});
-  const overrideRefs = useRef<Record<string, HTMLInputElement | null>>({});
-  const [previewTokens, setPreviewTokens] = useState<Record<string, string>>({
-    ...tokensByThemeState[initialTheme],
-    ...initialOverrides,
-  });
-  const debounceRef = useRef<number | null>(null);
-  const saveDebounceRef = useRef<number | null>(null);
-  const pendingPatchRef = useRef<{
-    overrides: Record<string, string>;
-    defaults: Record<string, string>;
-  }>({ overrides: {}, defaults: {} });
 
-  const mergedTokens = useMemo(
-    () => ({ ...tokensByThemeState[theme], ...overrides }),
-    [theme, tokensByThemeState, overrides]
-  );
+  const [contrastWarnings, setContrastWarnings] = useState<Record<string, string>>({});
+  const overrideRefs = useRef<Record<string, HTMLInputElement | null>>({});
 
   const textTokenKeys = useMemo(
     () =>
       Object.keys(tokensByThemeState[theme]).filter((k) =>
-        /text|foreground/i.test(k)
+        /text|foreground/i.test(k),
       ),
-    [theme, tokensByThemeState]
+    [theme, tokensByThemeState],
   );
 
   const bgTokenKeys = useMemo(
     () =>
       Object.keys(tokensByThemeState[theme]).filter((k) =>
-        /bg|background/i.test(k)
+        /bg|background/i.test(k),
       ),
-    [theme, tokensByThemeState]
+    [theme, tokensByThemeState],
   );
 
   const groupedTokens = useMemo(() => {
@@ -106,96 +100,11 @@ export function useThemeEditor({
     return groups;
   }, [theme, tokensByThemeState]);
 
-  const schedulePreviewUpdate = (tokens: Record<string, string>) => {
-    if (debounceRef.current) {
-      clearTimeout(debounceRef.current);
-    }
-    debounceRef.current = window.setTimeout(() => {
-      setPreviewTokens(tokens);
-      savePreviewTokens(tokens);
-    }, 100);
-  };
-
-  const scheduleSave = (
-    overridesPatch: Record<string, string>,
-    defaultsPatch: Record<string, string> = {}
-  ) => {
-    pendingPatchRef.current = {
-      overrides: {
-        ...pendingPatchRef.current.overrides,
-        ...overridesPatch,
-      },
-      defaults: {
-        ...pendingPatchRef.current.defaults,
-        ...defaultsPatch,
-      },
-    };
-    if (saveDebounceRef.current) {
-      clearTimeout(saveDebounceRef.current);
-    }
-    saveDebounceRef.current = window.setTimeout(async () => {
-      const payload = pendingPatchRef.current;
-      pendingPatchRef.current = { overrides: {}, defaults: {} };
-      try {
-        await patchShopTheme(shop, {
-          themeOverrides: payload.overrides,
-          themeDefaults: payload.defaults,
-        });
-      } catch (err) {
-        console.error(err);
-      }
-    }, 500);
-  };
-
-  // Persist merged tokens and broadcast to preview whenever they change
-  useEffect(() => {
-    schedulePreviewUpdate(mergedTokens);
-  }, [mergedTokens]);
-
   const handleWarningChange = (token: string, warning: string | null) => {
     setContrastWarnings((prev) => {
       const next = { ...prev };
       if (warning) next[token] = warning;
       else delete next[token];
-      return next;
-    });
-  };
-
-  const handleOverrideChange =
-    (key: string, defaultValue: string) => (value: string) => {
-      setOverrides((prev) => {
-        const next = { ...prev };
-        const patch: Record<string, string> = {};
-        if (!value || value === defaultValue) {
-          delete next[key];
-          patch[key] = defaultValue;
-        } else {
-          next[key] = value;
-          patch[key] = value;
-        }
-        scheduleSave(patch);
-        return next;
-      });
-    };
-
-  const handleReset = (key: string) => () => {
-    setOverrides((prev) => {
-      const next = { ...prev };
-      delete next[key];
-      scheduleSave({ [key]: tokensByThemeState[theme][key] });
-      return next;
-    });
-  };
-
-  const handleGroupReset = (keys: string[]) => () => {
-    setOverrides((prev) => {
-      const next = { ...prev };
-      const patch: Record<string, string> = {};
-      keys.forEach((k) => {
-        delete next[k];
-        patch[k] = tokensByThemeState[theme][k];
-      });
-      scheduleSave(patch);
       return next;
     });
   };
@@ -210,42 +119,6 @@ export function useThemeEditor({
         input.click();
       }
     }
-  };
-
-  useEffect(() => {
-    return () => {
-      if (debounceRef.current) {
-        clearTimeout(debounceRef.current);
-      }
-      if (saveDebounceRef.current) {
-        clearTimeout(saveDebounceRef.current);
-      }
-    };
-  }, []);
-
-  // Broadcast initial tokens so previews reflect the current theme on mount
-  useEffect(() => {
-    savePreviewTokens(previewTokens);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  const handleThemeChange = async (e: ChangeEvent<HTMLSelectElement>) => {
-    const newTheme = e.target.value;
-    setTheme(newTheme);
-    setOverrides({});
-    setThemeDefaults(tokensByThemeState[newTheme]);
-  };
-
-  const handleResetAll = async () => {
-    if (!window.confirm("Are you sure you want to reset all overrides?")) {
-      return;
-    }
-    const patch: Record<string, string> = {};
-    Object.keys(overrides).forEach((k) => {
-      patch[k] = tokensByThemeState[theme][k];
-    });
-    setOverrides({});
-    scheduleSave(patch);
   };
 
   return {
@@ -274,3 +147,4 @@ export function useThemeEditor({
     handleResetAll,
   };
 }
+
